+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rxweave/core
 
+## 0.5.4
+
+### Patch Changes
+
+- @rxweave/schema@0.5.4
+
 ## 0.5.3
 
 ### Patch Changes