+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rxweave/server
 
+## 0.5.4
+
+### Patch Changes
+
+- @rxweave/core@0.5.4
+- @rxweave/protocol@0.5.4
+- @rxweave/schema@0.5.4
+
 ## 0.5.3
 
 ### Patch Changes